requirements.txt
torch
torchvision
torchmetrics
opencv-python
Pillow
typer
pyyaml
ultralytics